Greatly improved smartFileToID logic
It now won't bother checking fileToID if the "fixed" file path is the same as
the original file path. It'll try to resolve relative components of the
directory path before attempting to find the realpath, and the parseM3U
function will try to resolve relative m3u paths before calling smartFileToID so
that it can resolve regardless of if the m3u was provided with an absolute path
or not.